perf(phai): lazy-load the embeddable RunViewer and unify the run-log loader
